refactor(home): extract blotted file into components
